Rename communicator to 'RestApiClient' and remove assembly attribute.
